Handle print and range functions
Python print is a very versatile function, which translates into the Rust macro print! or println!, which is also versatile but in different ways. We handle many of the translations, though not C-style format strings in Python, or formatted strings in Python.

Also range is very different. For example, range(100) in Python translates to 0..100 in Rust