Clean compile in Rust
Lots of changes to make the Rust sources clean(ish) compile. Some not very important warnings, like avoid "while true".

However, to achieve this, I had to cheat. The baselines are not the same as what the code generates. There is more work needed on adding "to_string" as required.