Handle type coercion of String/&str
In Python, all strings are type "str". In Rust we want strings passed as arguments to be "&str", but strings held in variables or returned to be "String". We need type coercion between these.

First, generate a dictionary that annotates the entire AST with the types of items at each node. (We do not care much about nodes whose types are irrelevant, such as operators or Load contexts.) Next, use this type info to determine whether we need to_string methods.