Support function calls
We support function calls within the same module, or in other modules on the sys.path. There are still quite a few caveats, especially to type inference of call return types. This only works if:

1. The function is in a separate module that can be loaded.
2. The function is invoked as module_name.function_name (no import as or from module import)
3. The function has type hints

Without these, the return type of the function is unknown. This may not matter, but may result in missing type coercions etc.