Partially fix binary operators on lists
Python allows you to write expressions like (a + b) * c where any or all of a, b and c are lists. The result is element-by-element arithmetic, using the smallest of the lists to define the resulting size.

Rust does not allow this. The reason is that binary operators in Rust would generate temporary lists, which would be horrifically inefficient. (The same is true in Python, but that is inefficient anyway.)

We can achieve the same thing in many cases by devectorising the arithmetic, which we do in Rust. (Python's numexpr library does much the same thing.)

This is not completely general. In particular, if some of the operands are not variables but function calls, we do not want to call the functions repeatedly. We leave this as a future issue.